# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI Service for processing board data with summarization and structurization capabilities",
        "title": "AIService API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Check if the service is running",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "description": "Get the status of a job by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Job"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{id}/abort": {
            "put": {
                "description": "Abort a job by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Abort a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/structurize": {
            "post": {
                "description": "Process a board and return a structured file hierarchy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Processing"
                ],
                "summary": "Structurize a board",
                "parameters": [
                    {
                        "description": "Structurize Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StructurizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StructurizeResponse"
                        }
                    },
                    "202": {
                        "description": "Job ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/summarize": {
            "post": {
                "description": "Process a board and return a summary of the content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Processing"
                ],
                "summary": "Summarize a board",
                "parameters": [
                    {
                        "description": "Summarize Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SummarizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SummarizeResponse"
                        }
                    },
                    "202": {
                        "description": "Job ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AnalyzeRequest": {
            "type": "object",
            "properties": {
                "requestType": {
                    "description": "summarize, structurize",
                    "type": "string"
                },
                "structurizeRequest": {
                    "$ref": "#/definitions/models.StructurizeRequest"
                },
                "summarizeRequest": {
                    "$ref": "#/definitions/models.SummarizeRequest"
                }
            }
        },
        "models.Board": {
            "type": "object",
            "properties": {
                "boardId": {
                    "type": "string"
                },
                "elements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Element"
                    }
                },
                "imageUrl": {
                    "type": "string"
                }
            }
        },
        "models.Element": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "inserted from text model",
                    "type": "string"
                },
                "cornerRadius": {
                    "description": "inserted from rectangle model",
                    "type": "integer"
                },
                "fill": {
                    "description": "цвет заливки",
                    "type": "string"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "points": {
                    "description": "[x, y], [x, y]",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "rotation": {
                    "type": "number"
                },
                "stroke": {
                    "description": "цвет обводки",
                    "type": "string"
                },
                "strokeWidth": {
                    "description": "толщина обводки",
                    "type": "integer"
                },
                "tension": {
                    "description": "давление",
                    "type": "number"
                },
                "type": {
                    "description": "rect, line, text, ellipse,",
                    "type": "string"
                },
                "width": {
                    "type": "number"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "models.File": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.File"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "main.go"
                },
                "type": {
                    "description": "doc, simple, graph,(поле children пустое) | section (содердит детей)",
                    "type": "string",
                    "example": "doc"
                }
            }
        },
        "models.Job": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "integer",
                    "format": "int64"
                },
                "id": {
                    "type": "string"
                },
                "request": {
                    "$ref": "#/definitions/models.AnalyzeRequest"
                },
                "retries": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.JobStatus"
                }
            }
        },
        "models.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "aborted"
            ],
            "x-enum-varnames": [
                "JobStatusPending",
                "JobStatusRunning",
                "JobStatusCompleted",
                "JobStatusFailed",
                "JobStatusAborted"
            ]
        },
        "models.StructurizeRequest": {
            "type": "object",
            "properties": {
                "board": {
                    "$ref": "#/definitions/models.Board"
                },
                "file": {
                    "$ref": "#/definitions/models.File"
                },
                "requestId": {
                    "type": "string"
                },
                "requestType": {
                    "description": "structurize",
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.StructurizeResponse": {
            "type": "object",
            "properties": {
                "aiTreeResponse": {
                    "description": "дерево ASCII файлов",
                    "type": "string"
                },
                "file": {
                    "$ref": "#/definitions/models.File"
                },
                "requestId": {
                    "type": "string"
                },
                "requestType": {
                    "description": "structurize",
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.SummarizeRequest": {
            "type": "object",
            "properties": {
                "board": {
                    "$ref": "#/definitions/models.Board"
                },
                "requestId": {
                    "type": "string"
                },
                "requestType": {
                    "description": "summarize",
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.SummarizeResponse": {
            "type": "object",
            "properties": {
                "requestId": {
                    "type": "string"
                },
                "requestType": {
                    "description": "summarize",
                    "type": "string"
                },
                "text": {
                    "description": "конкретный элемент - текст, который суммаризовал инфу по доске, расположенный в свободном пространстве доски",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Text"
                        }
                    ]
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.Text": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "fill": {
                    "description": "цвет заливки",
                    "type": "string"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "rotation": {
                    "type": "number"
                },
                "stroke": {
                    "description": "цвет обводки",
                    "type": "string"
                },
                "strokeWidth": {
                    "description": "толщина обводки",
                    "type": "integer"
                },
                "type": {
                    "description": "rect, line, text, ellipse,",
                    "type": "string"
                },
                "width": {
                    "type": "number"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        }
    }
}